Adding Python Jupiter Notebooks for velocity analysis
The structure I choose for the analysis looks like that: All stuff that runs in R is one script. This generates the .csv files which are used to prepare the AnnData objects. Then there is one Jupiter Notebook, which prepares the the Anndata objects and three others which are copied from the KappaVelo tutorial. I adjusted the scripts in a way that they work for us. To reproduce the analysis you have to run Notebook 2 and 3. Only remaining problem is that I have to find out how we can observe the velocities for individual genes, but this is rather a small issue and likely caused by annotation problems. KappaVelo Notebooks were copied from: https://github.com/HaghverdiLab/velocity_notebooks