major_update
align script and clean code to match most up-to-date figure layout. clean up results to match figure layout. align repo with figures as they currently stand for the manuscript.

Note three files are too large for github and along with other primary input and produced data will be moved to zenodo for hosting